Reuse KqpCommitLocks
