+++ ydb/core/tx/columnshard/transactions/operators/ev_write.cpp
@@ -5,23 +5,16 @@
 
 namespace NKikimr::NColumnShard {
 
-void KqpCommitLocks(ui64 origin, const NKikimrDataEvents::TKqpLocks* kqpLocks, NDataShard::TSysLocks& sysLocks) {
-    if (kqpLocks == nullptr) {
-        return;
+bool TEvWriteTransactionOperator::DoParse(TColumnShard& /*owner*/, const TString& data) {
+    NKikimrTxColumnShard::TCommitWriteTxBody commitTxBody;
+    if (!commitTxBody.ParseFromString(data)) {
+        return false;
     }
-
-    if (NDataShard::NeedCommitLocks(kqpLocks->GetOp())) {
-        // We assume locks have been validated earlier
-        for (const auto& lockProto : kqpLocks->GetLocks()) {
-            if (lockProto.GetDataShard() != origin) {
-                continue;
-            }
-            auto lockKey = NDataShard::MakeLockKey(lockProto);
-            sysLocks.CommitLock(lockKey);
-        }
-    } else {
-        NDataShard::KqpEraseLocks(origin, kqpLocks, sysLocks);
+    LockId = commitTxBody.GetLockId();
+    if (commitTxBody.HasKqpLocks() && !commitTxBody.GetKqpLocks().GetLocks().empty()) {
+        KqpLocks = commitTxBody.GetKqpLocks();
     }
+    return !!LockId;
 }
 
 TEvWriteTransactionOperator::TProposeResult TEvWriteTransactionOperator::DoStartProposeOnExecute(TColumnShard& owner, NTabletFlatExecutor::TTransactionContext& txc) {
@@ -31,7 +24,6 @@ TEvWriteTransactionOperator::TProposeResult TEvWriteTransactionOperator::DoStart
     return TProposeResult();
 }
 
-
 void TEvWriteTransactionOperator::SubscribeNewLocks(TColumnShard& owner, const TActorContext& ctx) {
     while (auto pendingSubscribeLock = owner.SysLocksTable().NextPendingSubscribeLock()) {
         ctx.Send(NLongTxService::MakeLongTxServiceID(owner.SelfId().NodeId()),
@@ -40,7 +32,6 @@ void TEvWriteTransactionOperator::SubscribeNewLocks(TColumnShard& owner, const T
                 pendingSubscribeLock.LockNodeId));
     }
 }
-
 
 bool TEvWriteTransactionOperator::ExecuteOnProgress(TColumnShard& owner, const NOlap::TSnapshot& version, NTabletFlatExecutor::TTransactionContext& txc) {
     auto operations = owner.OperationsManager->GetOperations(LockId);
@@ -53,47 +44,33 @@ bool TEvWriteTransactionOperator::ExecuteOnProgress(TColumnShard& owner, const N
         if (guardLocks.LockTxId) {
             switch (owner.SysLocksTable().EnsureCurrentLock()) {
                 case NDataShard::EEnsureCurrentLock::Success:
-                    // Lock is valid, we may continue with reads and side-effects
                     break;
-
                 case NDataShard::EEnsureCurrentLock::Broken:
-                    // Lock is valid, but broken, we could abort early in some
-                    // cases, but it doesn't affect correctness.
                     break;
-
                 case NDataShard::EEnsureCurrentLock::TooMany:
-                    // Lock cannot be created, it's not necessarily a problem
-                    // for read-only transactions, for non-readonly we need to
-                    // abort;
-                    return false; // TProposeResult(NKikimrTxColumnShard::EResultStatus::OVERLOADED, "too many locks");
+                    return false;
                 case NDataShard::EEnsureCurrentLock::Abort:
-                    return false; //TProposeResult(NKikimrTxColumnShard::EResultStatus::ERROR, "cannot acquire locks");
+                    return false;
             }
         }
 
         const auto tabletId = owner.TabletID();
         const auto* kqpLocks = &KqpLocks.value();
 
-        auto [validated, brokenLocks] = NDataShard::KqpValidateLocks(tabletId, owner.SysLocksTable(), kqpLocks, true, InReadSets);
+        bool validated = false;
+        std::tie(validated, BrokenLocks) = NDataShard::KqpValidateLocks(tabletId, owner.SysLocksTable(), kqpLocks, true, InReadSets);
         if (!validated) {
-            for (auto& brokenLock : brokenLocks) {
-                Y_UNUSED(brokenLock);
-             //   writeOp->GetWriteResult()->Record.MutableTxLocks()->Add()->Swap(&brokenLock);
-            }
-
-            KqpEraseLocks(tabletId, kqpLocks, owner.SysLocksTable());
+            NDataShard::KqpEraseLocks(tabletId, kqpLocks, owner.SysLocksTable());
             owner.SysLocksTable().ApplyLocks();
             // SubscribeNewLocks(ctx);
             if (locksDb.HasChanges()) {
                 //  op->SetWaitCompletionFlag(true);
-                return false; // EExecutionStatus::ExecutedNoMoreRestarts;
+                return false;
             }
-
-            // NKikimrDataEvents::TEvWriteResult::STATUS_LOCKS_BROKEN
-            return true; //TProposeResult(NKikimrTxColumnShard::EResultStatus::ERROR, "Operation is aborting because locks are not valid");
+            return true;
         }
 
-        KqpCommitLocks(tabletId, kqpLocks, owner.SysLocksTable()); // Do it on commit
+        NDataShard::KqpCommitLocks(tabletId, kqpLocks, owner.SysLocksTable(), [](const NKikimrDataEvents::TLock&){});
     }
 
     return owner.OperationsManager->CommitTransaction(owner, GetTxId(), txc, version);
