+++ ydb/core/tx/columnshard/transactions/operators/ev_write.h
@@ -36,17 +36,7 @@ namespace NKikimr::NColumnShard {
             ctx.Send(txInfo.Source, evResult.release(), 0, txInfo.Cookie);
         }
 
-        virtual bool DoParse(TColumnShard& /*owner*/, const TString& data) override {
-            NKikimrTxColumnShard::TCommitWriteTxBody commitTxBody;
-            if (!commitTxBody.ParseFromString(data)) {
-                return false;
-            }
-            LockId = commitTxBody.GetLockId();
-            if (commitTxBody.HasKqpLocks()) {
-                KqpLocks = commitTxBody.GetKqpLocks();
-            }
-            return !!LockId;
-        }
+        virtual bool DoParse(TColumnShard& /*owner*/, const TString& data) override;
 
     public:
         using TBase::TBase;
@@ -54,8 +44,16 @@ namespace NKikimr::NColumnShard {
         virtual bool ExecuteOnProgress(TColumnShard& owner, const NOlap::TSnapshot& version, NTabletFlatExecutor::TTransactionContext& txc) override;
 
         virtual bool CompleteOnProgress(TColumnShard& owner, const TActorContext& ctx) override {
-            auto result = NEvents::TDataEvents::TEvWriteResult::BuildCompleted(owner.TabletID(), GetTxId());
-            ctx.Send(TxInfo.Source, result.release(), 0, TxInfo.Cookie);
+            if (BrokenLocks.empty()) {
+                auto result = NEvents::TDataEvents::TEvWriteResult::BuildCompleted(owner.TabletID(), GetTxId());
+                ctx.Send(TxInfo.Source, result.release(), 0, TxInfo.Cookie);
+            } else {
+                auto result = NEvents::TDataEvents::TEvWriteResult::BuildError(owner.TabletID(), GetTxId(), NKikimrDataEvents::TEvWriteResult::STATUS_LOCKS_BROKEN, "Operation is aborting because locks are not valid");
+                for (auto& brokenLock : BrokenLocks) {
+                    result->Record.MutableTxLocks()->Add()->Swap(&brokenLock);
+                }
+                ctx.Send(TxInfo.Source, result.release(), 0, TxInfo.Cookie);
+            }
             return true;
         }
 
@@ -73,6 +71,8 @@ namespace NKikimr::NColumnShard {
         ui64 LockId = 0;
         std::optional<NKikimrDataEvents::TKqpLocks> KqpLocks;
         TInReadSets InReadSets;
+
+        TVector<NKikimrDataEvents::TLock> BrokenLocks;
     };
 
 }
